Split practice-view.test.tsx by interaction surface
